refactor: change namespace name
