+++ Conflux.API.Tests/Controllers/ProjectsControllerTests.cs
@@ -9,6 +9,7 @@ using System.Text.Json;
 using Conflux.Data;
 using Conflux.Domain;
 using Conflux.Domain.Logic.DTOs;
+using Conflux.Domain.Logic.DTOs.Patch;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit;
