+++ Conflux.API/Controllers/ContributorsController.cs
@@ -5,6 +5,7 @@
 
 using Conflux.Domain;
 using Conflux.Domain.Logic.DTOs;
+using Conflux.Domain.Logic.DTOs.Patch;
 using Conflux.Domain.Logic.Services;
 using Microsoft.AspNetCore.Mvc;
