+++ Conflux.Domain.Logic/Services/ContributorsService.cs
@@ -5,6 +5,7 @@
 
 using Conflux.Data;
 using Conflux.Domain.Logic.DTOs;
+using Conflux.Domain.Logic.DTOs.Patch;
 using Conflux.Domain.Logic.Exceptions;
 using Microsoft.EntityFrameworkCore;
